Avoid ghosts in the list of installed apps
When we run a system app, we create the org.foo.App/data directory
in the user location. This was causing the app to show up in the
list of user apps, when not using --show-details. Avoid this by
always listing full references, and only hiding the arch and branch
information without --show-details.