Support filtering of session bus
If the session bus is allowed we get unfiltered access to it, but if
it is not, then we install a filter with the policy specified by the
runtime and the app.

By default you can always own your app-id and wildcarded names below
it (i.e. org.gnome.gedit and org.gnome.gedit.*). On top of that we add
all the policies specified in the metadata for the runtime, and then
the metadata from the application.

The metadata policy config looks like this:

[Session Bus Policy]
ca.desrt.dconf=talk
org.gtk.*=see
org.my.name=own