libcontainer: Pare down to just "run in root" API
rpm-ostree is going through some awkwardness trying to
support being run both inside and outside of a container.

For now, let's drop all recursive container usage.  There is still
some value in readding this in the future - for example, downloading
packages requires networking, but `%post` scripts do not.  But we
should really solve that when we return to running unprivileged or the
like.