autogen.sh: rely on VPATH when generating Makefile-*.am.inc
This avoids hard-coding the (potentially absolute) srcdir used at
autogen time into the build system. It's fine for an Automake build
system to use relative paths in dependencies and rely on them being
interpreted as relative to ${srcdir} even though the current working
directory is ${buildddir}, because of make's "VPATH" feature, and we
do it all the time in the handwritten parts of the build system.

In particular, this (finally) makes distcheck work.

Signed-off-by: Simon McVittie <smcv@debian.org>