Install/update/uninstall related refs
Any "related" references from the same remote are automatically
handled when installing, unless you specify --no-related.

The set of related refs is based on the declared extensions for
the main ref. By default, all such extensions are automatically
installed/updated unless they have "no-autodownload" set to false
in the extension info, and its not already installed locally.

Only extensions marked as "autodelete" will be automatically deleted
when you remove the main ref.

There is some special handling:

Any *.Debug or *.Locale extension is automatically treated as "autodelete".

A *.Locale extension that is autodownloaded will be subsetted using subpaths
such that existing subpaths are kept, and the current locale is added.