Make session helper manage SELinux labels
Add RequestSelinuxLabel and ReleaseSelinuxLabel methods to
the flatpak session helper. The code behind these is doing
essentially the same as docker-daemon does for containers:
Use the type part of the lxc exec and file contexts, and
make them unique by appending mcs levels.

We store the labels in a flatpak db table, so we can hand
out the same labels to an app as long as it is installed.

Currently, we don't clean up the labels at all, which
should not be a problem in the short term.