When installing to the system dir, don't consider --user dependencies
It makes no sense to allow you to install an application systemwide
if it only works for your user. However, its generally fine if to
install a per-user app that relies on a system-installed runtime.

In particular, this fixes installing system-wide extra-data apps where
we need the runtime to be system-installed to run the unpack script.